Delete docker-compose.yml symlink
Fix #67

A symlink to a docker-compose.yml file was added to the project but
should be removed. Because the link points to a location that doesn't
exist on a local dev's hard drive, the file always shows up as being
modified -- preventing operations such as rebase. There's no need to
add a symlink to a git repo in any case.

Signed-off-by: Trevor Conn <trevor_conn@dell.com>