core/certs.go, kongParameters.go: use JSON instead of form
Use JSON to post the certificates rather than the Form, as JSON is now
more reliable than using the application/x-www-form-urlencoded method.

There are some race conditions which sometimes lead to Kong not picking
up the SNIs setting for the certificate, which in turn leads to Kong
picking the wrong certs to serve with for endpoints.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>